feat: refactored AI gateway orchestration harness with agents, state management, and telemetry support
